requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
crispy-tailwind==0.5.0
distlib==0.3.3
Django==3.2.9
django-crispy-forms==1.13.0
filelock==3.3.1
platformdirs==2.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
virtualenv==20.8.1
whitenoise==5.3.0
xlwt==1.3.0
django-tinymce